add benchmark for charconv_hexfloat_float and hexfloat_float.cc
fixing dragonbox, thanks to jk-jeon

+++ benchmark/0004.file_io_floating/charconv_hexfloat_float.cc
@@ -9,8 +9,8 @@ int main()
 {
 	constexpr std::size_t N{10000000};
 	std::mt19937_64 eng{};
-	std::uniform_real_distribution<double> dis(DBL_MIN,DBL_MAX);
-	std::vector<double> vec;
+	std::uniform_real_distribution<float> dis(FLT_MIN,FLT_MAX);
+	std::vector<float> vec;
 	vec.reserve(N);
 	for(std::size_t i{};i!=N;++i)
 	{
